Update README.md for Better User Onboarding
This commit enhances the README document to make it more informative and user-friendly. It adds a more detailed introduction, steps to add Ozone to a channel, instructions to use Ozone as a Limnoria plugin, and a Troubleshooting section to address potential issues. It also updates the Code of Conduct and Licensing sections for improved visibility and comprehension.